resolve test script

+++ src/models/person.model.js
@@ -2,19 +2,20 @@
 const Person = require("./person.mongo");
 
 //To set the default serial number...
-const defaultSerialNum = 1;
+// const defaultSerialNum = 1;
 
 //To get the current serial number...
-async function getSerialNum() {
-  const existingSerialNum = Person.find({})
+async function getExistingNumberOfPersons() {
+  try {
+    const existingNumberOfPersons = Person.find({})
     .then((people) => {
       return people.length;
     })
-    .catch((err) => {
-      console.error(err);
-      return {error : 'Error getting serial number'}
-    });
-  return existingSerialNum;
+    return existingNumberOfPersons;
+  } catch (error) {
+    console.error(error)
+    return {error : 'error getting serial number'}
+  }
 }
 
 //To get all saved persons from the database...
@@ -23,7 +24,7 @@ async function getPersonData() {
     return Person.find({}, { _id: 0, __v: 0 }).sort({ serialNum: 1 });
   } catch (error) {
     console.error(error);
-    return {error : 'Error getting person data'}
+    return {error : 'error getting all person data'}
   }
 }
 
@@ -32,55 +33,58 @@ async function getSinglePersonData(personParam) {
   try {
     return isNaN(personParam)
     ? Person.findOne({ name: personParam }, { _id: 0, __v: 0 })
-    : Person.findOne({ serialNum: personParam }, { _id: 0, __v: 0 });
+    : Person.findOne({ id: personParam }, { _id: 0, __v: 0 });
   } catch (error) {
     console.error(error);
-    return {error : 'Error getting person data'};
+    return {error : 'error getting person data'};
   }
 }
 
 //To add a person to the database...
 async function addPersonData(person) {
   try {
-    const newSerialNum = (await getSerialNum()) + 1;
-    if(isNaN(newSerialNum)) return {error : 'Error getting serial number'};
+    const newUser_id = ((await getExistingNumberOfPersons()) + 1);
     const newPerson = Object.assign(person, {
-      serialNum: newSerialNum,
+      id: newUser_id,
     });
     await Person.create(newPerson);
     return newPerson;  
   } catch (error) {
     console.error(error);
-    return {error : 'Error in adding person data'};
+    return {error : 'error in adding person data'};
   }
 }
 
 //To update a person in the database...
-async function updatePersonData(queryUser_id, updatedPerson) {
+async function updatePersonData(queryUser_id, updatePerson) {
   try {
+    const existingPerson = await getSinglePersonData(queryUser_id);
+    if(!existingPerson){ return {ok: false}}
     await Person.findOneAndUpdate(
       {
-        serialNum: queryUser_id,
+        id: existingPerson.id,
       },
-      updatedPerson,
+      updatePerson,
       {
         upsert: true,
       }
     );
-    return updatedPerson;  
+    return {ok: true};  
   } catch (error) {
     console.error(error);
-    return {error : 'Error occured in updating person data'};
+    return {error : 'error occured in updating person data'};
   }
 }
 
 //To delete a person from the database...
 async function deletePersonData(queryUser_id) {
   try {
+    const existingPerson = await getSinglePersonData(queryUser_id);
+    if(!existingPerson){ return { ok: false}}
     await Person.findOneAndDelete({
-      serialNum: queryUser_id,
+      id: existingPerson.id,
     });
-    return true;  
+    return {ok: true};  
   } catch (error) {
     console.error(error);
     return {error : 'error occured in deleting person data'};
